- missed file changes in wmsk version upgrade

+++ Assets/WorldMapStrategyKit/Demos/Path Finding Examples/204 PathFinding by Cells/DemoCellsPathFinding.cs
@@ -10,7 +10,7 @@ namespace WorldMapStrategyKit {
 		int startCellIndex;
 		GameObjectAnimator tank;
 		List<int> path;
-		int pathCost;
+		float pathCost;
 		List<TextMesh> texts = new List<TextMesh> ();
 
 		void Start () {
@@ -184,7 +184,7 @@ namespace WorldMapStrategyKit {
 				map.SetCellTemporaryColor (cellIndex, pathColor);
 
 				// Show the accumulated cost
-				int accumCost = map.GetCellPathCost (cellIndex);
+				float accumCost = map.GetCellPathCost (cellIndex);
 				Vector3 cellPosition = map.GetCellPosition (cellIndex);
 				TextMesh text = map.AddMarker2DText (accumCost.ToString (), cellPosition);
 				text.transform.localScale *= 0.3f;	// make font smaller
